+++ src/Sensorario/ValueObject/Exceptions/InvalidValueException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Sensorario\ValueObject\Exceptions;
+
+use Exception;
+
+final class InvalidValueException extends Exception { }